[DEV-3685] Update migration
